+++ contrib/babelfishpg_tsql/sql/upgrades/babelfishpg_tsql--4.4.0--5.0.0.sql
@@ -7600,8 +7600,8 @@ BEGIN
        (p_hour::SMALLINT NOT BETWEEN 0 AND 23) OR
        (p_minute::SMALLINT NOT BETWEEN 0 AND 59) OR
        (p_seconds::SMALLINT NOT BETWEEN 0 AND 59) OR
-       (p_fractions::SMALLINT NOT BETWEEN 0 AND 9999999) OR
-       (p_fractions::SMALLINT != 0 AND char_length(v_fractions) > v_precision))
+       (p_fractions::INT NOT BETWEEN 0 AND 9999999) OR
+       (p_fractions::INT != 0 AND char_length(v_fractions) > v_precision))
    THEN
       RAISE invalid_datetime_format;
    ELSIF (v_precision NOT BETWEEN 0 AND 7) THEN
